napi

+++ crates/canvas/src/dom/point.rs
@@ -1,21 +1,21 @@
 #[napi]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Serialize, Deserialize)]
 pub struct DomPoint {
-    pub w: i32,
-    pub x: i32,
-    pub y: i32,
-    pub z: i32,
+    pub w: f64,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 #[napi]
 impl DomPoint {
     #[napi(constructor)]
-    pub fn new(w: Option<i32>, x: Option<i32>, y: Option<i32>, z: Option<i32>) -> Self {
+    pub fn new(w: Option<f64>, x: Option<f64>, y: Option<f64>, z: Option<f64>) -> Self {
         Self {
-            w: w.unwrap_or(0),
-            x: x.unwrap_or(0),
-            y: y.unwrap_or(0),
-            z: z.unwrap_or(0),
+            w: w.unwrap_or(0.0),
+            x: x.unwrap_or(0.0),
+            y: y.unwrap_or(0.0),
+            z: z.unwrap_or(0.0),
         }
     }
 }
